add areas
