+++ app/routes/organization/$slug/settings/general.tsx
@@ -16,8 +16,10 @@ import InputAdd from "~/components/FormElements/InputAdd/InputAdd";
 import InputText from "~/components/FormElements/InputText/InputText";
 import SelectAdd from "~/components/FormElements/SelectAdd/SelectAdd";
 import TextAreaWithCounter from "~/components/FormElements/TextAreaWithCounter/TextAreaWithCounter";
-import { createAreaOptionFromData } from "~/lib/profile/createAreaOptionFromData";
-import { objectListOperationResolver } from "~/lib/utils/components";
+import {
+  createAreaOptionFromData,
+  objectListOperationResolver,
+} from "~/lib/utils/components";
 import { socialMediaServices } from "~/lib/utils/socialMediaServices";
 import {
   FormError,
