+++ app/routes/profile/$username/settings/general.tsx
@@ -1,4 +1,4 @@
-import * as React from "react";
+import React from "react";
 import { FormProvider, useForm } from "react-hook-form";
 import {
   ActionFunction,
@@ -18,8 +18,10 @@ import SelectAdd from "~/components/FormElements/SelectAdd/SelectAdd";
 import SelectField from "~/components/FormElements/SelectField/SelectField";
 import TextAreaWithCounter from "~/components/FormElements/TextAreaWithCounter/TextAreaWithCounter";
 import useCSRF from "~/lib/hooks/useCSRF";
-import { createAreaOptionFromData } from "~/lib/profile/createAreaOptionFromData";
-import { objectListOperationResolver } from "~/lib/utils/components";
+import {
+  createAreaOptionFromData,
+  objectListOperationResolver,
+} from "~/lib/utils/components";
 import { socialMediaServices } from "~/lib/utils/socialMediaServices";
 import {
   FormError,
